+++ docs/html/search/classes_1.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['datestrings_270',['DateStrings',['../classace__time_1_1DateStrings.html',1,'ace_time']]],
-  ['datetuple_271',['DateTuple',['../structace__time_1_1extended_1_1DateTuple.html',1,'ace_time::extended']]]
+  ['datestrings_282',['DateStrings',['../classace__time_1_1DateStrings.html',1,'ace_time']]],
+  ['datetuple_283',['DateTuple',['../structace__time_1_1extended_1_1DateTuple.html',1,'ace_time::extended']]]
 ];