+++ docs/html/search/classes_3.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['linkentrybroker_279',['LinkEntryBroker',['../classace__time_1_1basic_1_1LinkEntryBroker.html',1,'ace_time::basic::LinkEntryBroker'],['../classace__time_1_1extended_1_1LinkEntryBroker.html',1,'ace_time::extended::LinkEntryBroker']]],
-  ['linkmanager_280',['LinkManager',['../classace__time_1_1LinkManager.html',1,'ace_time']]],
-  ['linkregistrar_281',['LinkRegistrar',['../classace__time_1_1basic_1_1LinkRegistrar.html',1,'ace_time::basic::LinkRegistrar'],['../classace__time_1_1extended_1_1LinkRegistrar.html',1,'ace_time::extended::LinkRegistrar']]],
-  ['linkregistrartemplate_282',['LinkRegistrarTemplate',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrartemplate_3c_20basic_3a_3alinkentry_2c_20basic_3a_3alinkentrybroker_2c_20basic_3a_3alinkregistrybroker_20_3e_283',['LinkRegistrarTemplate&lt; basic::LinkEntry, basic::LinkEntryBroker, basic::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrartemplate_3c_20extended_3a_3alinkentry_2c_20extended_3a_3alinkentrybroker_2c_20extended_3a_3alinkregistrybroker_20_3e_284',['LinkRegistrarTemplate&lt; extended::LinkEntry, extended::LinkEntryBroker, extended::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
-  ['linkregistrybroker_285',['LinkRegistryBroker',['../classace__time_1_1basic_1_1LinkRegistryBroker.html',1,'ace_time::basic::LinkRegistryBroker'],['../classace__time_1_1extended_1_1LinkRegistryBroker.html',1,'ace_time::extended::LinkRegistryBroker']]],
-  ['localdate_286',['LocalDate',['../classace__time_1_1LocalDate.html',1,'ace_time']]],
-  ['localdatetime_287',['LocalDateTime',['../classace__time_1_1LocalDateTime.html',1,'ace_time']]],
-  ['localtime_288',['LocalTime',['../classace__time_1_1LocalTime.html',1,'ace_time']]]
+  ['linkentrybroker_291',['LinkEntryBroker',['../classace__time_1_1basic_1_1LinkEntryBroker.html',1,'ace_time::basic::LinkEntryBroker'],['../classace__time_1_1extended_1_1LinkEntryBroker.html',1,'ace_time::extended::LinkEntryBroker']]],
+  ['linkmanager_292',['LinkManager',['../classace__time_1_1LinkManager.html',1,'ace_time']]],
+  ['linkregistrar_293',['LinkRegistrar',['../classace__time_1_1basic_1_1LinkRegistrar.html',1,'ace_time::basic::LinkRegistrar'],['../classace__time_1_1extended_1_1LinkRegistrar.html',1,'ace_time::extended::LinkRegistrar']]],
+  ['linkregistrartemplate_294',['LinkRegistrarTemplate',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrartemplate_3c_20basic_3a_3alinkentry_2c_20basic_3a_3alinkentrybroker_2c_20basic_3a_3alinkregistrybroker_20_3e_295',['LinkRegistrarTemplate&lt; basic::LinkEntry, basic::LinkEntryBroker, basic::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrartemplate_3c_20extended_3a_3alinkentry_2c_20extended_3a_3alinkentrybroker_2c_20extended_3a_3alinkregistrybroker_20_3e_296',['LinkRegistrarTemplate&lt; extended::LinkEntry, extended::LinkEntryBroker, extended::LinkRegistryBroker &gt;',['../classace__time_1_1internal_1_1LinkRegistrarTemplate.html',1,'ace_time::internal']]],
+  ['linkregistrybroker_297',['LinkRegistryBroker',['../classace__time_1_1basic_1_1LinkRegistryBroker.html',1,'ace_time::basic::LinkRegistryBroker'],['../classace__time_1_1extended_1_1LinkRegistryBroker.html',1,'ace_time::extended::LinkRegistryBroker']]],
+  ['localdate_298',['LocalDate',['../classace__time_1_1LocalDate.html',1,'ace_time']]],
+  ['localdatetime_299',['LocalDateTime',['../classace__time_1_1LocalDateTime.html',1,'ace_time']]],
+  ['localtime_300',['LocalTime',['../classace__time_1_1LocalTime.html',1,'ace_time']]]
 ];