+++ docs/html/search/classes_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['yearmonthtuple_304',['YearMonthTuple',['../structace__time_1_1extended_1_1YearMonthTuple.html',1,'ace_time::extended']]]
+  ['yearmonthtuple_316',['YearMonthTuple',['../structace__time_1_1extended_1_1YearMonthTuple.html',1,'ace_time::extended']]]
 ];