+++ examples/CompareAceTimeToHinnantDate/BenchmarkAceTime.h
@@ -1,0 +1,12 @@
+#ifndef BENCHMARK_ACE_TIME_H
+#define BENCHMARK_ACE_TIME_H
+
+#include <string>
+#include <vector>
+
+extern void benchmarkAceTime(
+    const std::vector<std::string>& zones,
+    int startYear,
+    int untilYear);
+
+#endif