+++ src/ace_time/ExtendedZone.h
@@ -6,6 +6,7 @@
 #ifndef ACE_TIME_EXTENDED_ZONE_H
 #define ACE_TIME_EXTENDED_ZONE_H
 
+#include <AceCommon.h> // KString
 #include "internal/ZoneInfo.h"
 #include "internal/ExtendedBrokers.h"
 
@@ -36,6 +37,13 @@ class ExtendedZone {
     ExtendedZone(const extended::ZoneInfoBroker& zoneInfo):
         mZoneInfoBroker(zoneInfo) {}
 
+    // Use default copy constructor and assignment operator
+    ExtendedZone(const ExtendedZone&) = default;
+    ExtendedZone& operator=(const ExtendedZone&) = default;
+
+    /** Return true if zoneInfo is null. */
+    bool isNull() const { return mZoneInfoBroker.isNull(); }
+
     /** Print the full zone name to printer. Example "America/Los_Angeles". */
     void printNameTo(Print& printer) const;
 
@@ -51,11 +59,22 @@ class ExtendedZone {
       return mZoneInfoBroker.zoneId();
     }
 
-  private:
-    // disable copy constructor and assignment operator
-    ExtendedZone(const ExtendedZone&) = delete;
-    ExtendedZone& operator=(const ExtendedZone&) = delete;
+    /** Return the STDOFF of the last ZoneEra. */
+    int16_t stdOffsetMinutes() const {
+      uint8_t numEras = mZoneInfoBroker.numEras();
+      extended::ZoneEraBroker zeb = mZoneInfoBroker.era(numEras - 1);
+      return zeb.offsetMinutes();
+    }
 
+    /** Return the name as a KString. */
+    ace_common::KString kname() const {
+      const auto* name = isNull() ? nullptr : mZoneInfoBroker.name();
+      const internal::ZoneContext* zoneContext = mZoneInfoBroker.zoneContext();
+      return ace_common::KString(
+          name, zoneContext->fragments, zoneContext->numFragments);
+    }
+
+  private:
     const extended::ZoneInfoBroker mZoneInfoBroker;
 };