+++ src/ace_time/ZoneProcessorCache.h
@@ -15,20 +15,26 @@
 namespace ace_time {
 
 /**
- * A cache of ZoneProcessors that provides a ZoneProcessor to the TimeZone upon
- * request by the ZoneManager.
- *
- * @tparam SIZE number of zone processors, should be approximate the number
- *    zones *concurrently* used in the app. It is expected that this will be
- *    small. It can be 1 if the app never changes the TimeZone. It should be 2
- *    if the user is able to select different timezones from a menu.
- * @tparam ZP type of ZoneProcessor (BasicZoneProcessor or
- *    ExtendedZoneProcessor)
+ * The template class of BasicZoneProcessorCacheBase or
+ * ExtendedZoneProcessorCacheBase. The common implementation
+ * BasicZoneProcessorCache<SIZE> and ExtendedZoneProcessorCacheBase<SIZE>
+ * creates the caches inside the class itself, which will normally be created at
+ * static initialization time. An alternative implementation would create the
+ * cache on the heap.
  */
-template<uint8_t SIZE, typename ZP>
-class ZoneProcessorCacheTemplate {
+template <typename ZP>
+class ZoneProcessorCacheBaseTemplate {
   public:
-    ZoneProcessorCacheTemplate() = default;
+    ZoneProcessorCacheBaseTemplate(ZP* zoneProcessors, uint8_t size) :
+      mSize(size),
+      mZoneProcessors(zoneProcessors)
+    {}
+
+    /** Return the size of the cache. */
+    uint8_t size() const { return mSize; }
+
+    /** Get the ZoneProcessor at index i. */
+    ZP* getZoneProcessorAtIndex(uint8_t i) { return &mZoneProcessors[i]; }
 
     /**
      * Get ZoneProcessor from either a ZoneKey, either a basic::ZoneInfo or an
@@ -41,29 +47,25 @@ class ZoneProcessorCacheTemplate {
       // Allocate the next ZoneProcessor in the cache using round-robin.
       zoneProcessor = &mZoneProcessors[mCurrentIndex];
       mCurrentIndex++;
-      if (mCurrentIndex >= SIZE) mCurrentIndex = 0;
+      if (mCurrentIndex >= mSize) mCurrentIndex = 0;
       zoneProcessor->setZoneKey(zoneKey);
       return zoneProcessor;
     }
 
-    /** Return the ZoneProcessor at position i. Used for initialization. */
-    ZP* getZoneProcessor(uint8_t i) {
-      return &mZoneProcessors[i];
-    }
-
   private:
     // disable copy constructor and assignment operator
-    ZoneProcessorCacheTemplate(const ZoneProcessorCacheTemplate&) = delete;
-    ZoneProcessorCacheTemplate& operator=(const ZoneProcessorCacheTemplate&)
+    ZoneProcessorCacheBaseTemplate(const ZoneProcessorCacheBaseTemplate&)
         = delete;
+    ZoneProcessorCacheBaseTemplate& operator=(
+        const ZoneProcessorCacheBaseTemplate&) = delete;
 
     /**
      * Find an existing ZoneProcessor with the ZoneInfo given by zoneInfoKey.
      * Returns nullptr if not found. This is a linear search, which should
-     * be perfectly ok if SIZE is small, say <= 5.
+     * be perfectly ok if mSize is small, say <= 5.
      */
     ZP* findUsingZoneKey(uintptr_t zoneKey) {
-      for (uint8_t i = 0; i < SIZE; i++) {
+      for (uint8_t i = 0; i < mSize; i++) {
         ZP* zoneProcessor = &mZoneProcessors[i];
         if (zoneProcessor->equalsZoneKey(zoneKey)) {
           return zoneProcessor;
@@ -72,19 +74,78 @@ class ZoneProcessorCacheTemplate {
       return nullptr;
     }
 
+  private:
+    uint8_t const mSize;
     uint8_t mCurrentIndex = 0;
-    ZP mZoneProcessors[SIZE];
+    ZP* const mZoneProcessors;
 };
+
+/**
+ * Base class for all ZoneProcessorCache implementations that use a
+ * BasicZoneProcessor.
+ */
+using BasicZoneProcessorCacheBase =
+    ZoneProcessorCacheBaseTemplate<BasicZoneProcessor>;
+
+/**
+ * Base class for all ZoneProcessorCache implementations that use an
+ * ExtendedZoneProcessor.
+ */
+using ExtendedZoneProcessorCacheBase =
+    ZoneProcessorCacheBaseTemplate<ExtendedZoneProcessor>;
 
 #if 1
-template<uint8_t SIZE>
-class BasicZoneProcessorCache: public ZoneProcessorCacheTemplate<
-    SIZE, BasicZoneProcessor> {
+/**
+ * An implementation of a BasicZoneProcessorCacheBase where the cache of size
+ * SIZE is embedded into the class itself. This is expected to be created as a
+ * global object and passed into the BasicZoneManager.
+ *
+ * @tparam SIZE number of zone processors, should be approximate the number
+ *    zones *concurrently* used in the app. It is expected that this will be
+ *    small. It can be 1 if the app never changes the TimeZone. It should be 2
+ *    if the user is able to select different timezones from a menu.
+ */
+template <uint8_t SIZE>
+class BasicZoneProcessorCache : public BasicZoneProcessorCacheBase {
+  public:
+    BasicZoneProcessorCache() :
+      BasicZoneProcessorCacheBase(mZoneProcessors, SIZE)
+    {}
+
+  private:
+    // disable copy constructor and assignment operator
+    BasicZoneProcessorCache(const BasicZoneProcessorCache&) = delete;
+    BasicZoneProcessorCache& operator=(const BasicZoneProcessorCache&) = delete;
+
+  private:
+    BasicZoneProcessor mZoneProcessors[SIZE];
 };
 
-template<uint8_t SIZE>
-class ExtendedZoneProcessorCache: public ZoneProcessorCacheTemplate<
-    SIZE, ExtendedZoneProcessor> {
+/**
+ * An implementation of an ExtendedZoneProcessorCacheBase where the cache of
+ * size SIZE is embedded into the class itself. This is expected to be created
+ * as a global object and passed into the ExtendedZoneManager.
+ *
+ * @tparam SIZE number of zone processors, should be approximate the number
+ *    zones *concurrently* used in the app. It is expected that this will be
+ *    small. It can be 1 if the app never changes the TimeZone. It should be 2
+ *    if the user is able to select different timezones from a menu.
+ */
+template <uint8_t SIZE>
+class ExtendedZoneProcessorCache: public ExtendedZoneProcessorCacheBase {
+  public:
+    ExtendedZoneProcessorCache() :
+      ExtendedZoneProcessorCacheBase(mZoneProcessors, SIZE)
+    {}
+
+  private:
+    // disable copy constructor and assignment operator
+    ExtendedZoneProcessorCache(const ExtendedZoneProcessorCache&) = delete;
+    ExtendedZoneProcessorCache& operator=(const ExtendedZoneProcessorCache&)
+        = delete;
+
+  private:
+    ExtendedZoneProcessor mZoneProcessors[SIZE];
 };
 #else
 
@@ -94,11 +155,11 @@ class ExtendedZoneProcessorCache: public ZoneProcessorCacheTemplate<
 // seems to be no difference in code size between the two. The compiler seems
 // to optimize away the vtables of the parent and child classes.
 
-template<uint8_t SIZE>
+template <uint8_t SIZE>
 using BasicZoneProcessorCache = ZoneProcessorCacheTemplate<
     SIZE, BasicZoneProcessor>;
 
-template<uint8_t SIZE>
+template <uint8_t SIZE>
 using ExtendedZoneProcessorCache  = ZoneProcessorCacheTemplate<
     SIZE, ExtendedZoneProcessor>;
 #endif